#324 move redux stuff to own module
