+++ client/lib/common/src/data/kio/kio-actions.js
@@ -2,7 +2,7 @@
 import _ from 'lodash';
 import {createAction} from 'redux-actions';
 import Type from './kio-types';
-import {flummoxCompatWrap} from 'common/src/util';
+import {flummoxCompatWrap} from 'common/src/redux-middlewares';
 import request from 'common/src/superagent';
 import {Services} from 'common/src/data/services';
 import {Provider, RequestConfig, saveRoute} from 'common/src/oauth-provider';
